requirements.txt
pycairo==1.10.0
numpy==1.11.1
scipy==0.18.1
Pillow==4.0.0
